Final fixes

+++ alx_book_store.sql
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
     orderdetailid INT PRIMARY KEY,
     order_id INT,
     book_id INT,
-    quantity INT,
+    quantity DOUBLE,
     CONSTRAINT fk_order
         FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     CONSTRAINT fk_book
